+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/CompiledExpression.kt
@@ -1,0 +1,16 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+import dev.sargunv.maplibrecompose.expression.ExpressionContext
+import dev.sargunv.maplibrecompose.expression.value.ExpressionValue
+
+/**
+ * An [Expression] reduced to only those data types supported by the MapLibre SDKs. This can be
+ * thought of as an intermediate representation between the high level expression DSL and the
+ * platform-specific encoding.
+ */
+public sealed interface CompiledExpression<out T : ExpressionValue> : Expression<T> {
+  override fun compile(context: ExpressionContext): CompiledExpression<T> = this
+
+  @Suppress("UNCHECKED_CAST")
+  override fun <X : ExpressionValue> cast(): CompiledExpression<X> = this as CompiledExpression<X>
+}
