+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/CompiledListLiteral.kt
@@ -1,0 +1,20 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+import dev.sargunv.maplibrecompose.expression.value.ExpressionValue
+import dev.sargunv.maplibrecompose.expression.value.ListValue
+
+/** A [Literal] representing a JSON array with elements all [CompiledLiteral]. */
+public data class CompiledListLiteral<T : ExpressionValue>
+private constructor(override val value: List<CompiledLiteral<T, *>>) :
+  CompiledLiteral<ListValue<T>, List<Literal<T, *>>> {
+  override fun visit(block: (Expression<*>) -> Unit) {
+    block(this)
+    value.forEach { it.visit(block) }
+  }
+
+  public companion object {
+    internal fun <T : ExpressionValue> of(
+      value: List<CompiledLiteral<T, *>>
+    ): CompiledListLiteral<T> = CompiledListLiteral(value)
+  }
+}
