+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/DpOffsetLiteral.kt
@@ -1,0 +1,22 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.unit.DpOffset
+import dev.sargunv.maplibrecompose.expression.ExpressionContext
+import dev.sargunv.maplibrecompose.expression.value.DpOffsetValue
+
+/** A [Literal] representing a [DpOffset] value. */
+public data class DpOffsetLiteral private constructor(override val value: DpOffset) :
+  Literal<DpOffsetValue, DpOffset> {
+  override fun compile(context: ExpressionContext): CompiledLiteral<DpOffsetValue, Offset> =
+    OffsetLiteral.of(Offset(value.x.value, value.y.value)).cast()
+
+  override fun visit(block: (Expression<*>) -> Unit): Unit = block(this)
+
+  public companion object {
+    private val zero = DpOffsetLiteral(DpOffset.Zero)
+
+    public fun of(value: DpOffset): DpOffsetLiteral =
+      if (value == DpOffset.Zero) zero else DpOffsetLiteral(value)
+  }
+}
