+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/DpPaddingLiteral.kt
@@ -1,0 +1,18 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+import androidx.compose.foundation.layout.PaddingValues
+import dev.sargunv.maplibrecompose.expression.ZeroPadding
+import dev.sargunv.maplibrecompose.expression.value.DpPaddingValue
+
+/** A [Literal] representing a [PaddingValues] value. */
+public data class DpPaddingLiteral private constructor(override val value: PaddingValues.Absolute) :
+  CompiledLiteral<DpPaddingValue, PaddingValues.Absolute> {
+  override fun visit(block: (Expression<*>) -> Unit): Unit = block(this)
+
+  public companion object {
+    private val zero = DpPaddingLiteral(ZeroPadding)
+
+    public fun of(value: PaddingValues.Absolute): DpPaddingLiteral =
+      if (value == ZeroPadding) zero else DpPaddingLiteral(value)
+  }
+}
