+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/EnumLiteral.kt
@@ -1,0 +1,17 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+import dev.sargunv.maplibrecompose.expression.ExpressionContext
+import dev.sargunv.maplibrecompose.expression.value.EnumValue
+
+/** A [Literal] representing an enum value of type [T]. */
+public data class EnumLiteral<T : EnumValue<T>> private constructor(override val value: T) :
+  Literal<T, T> {
+  override fun compile(context: ExpressionContext): CompiledLiteral<T, String> =
+    value.literal.cast()
+
+  override fun visit(block: (Expression<*>) -> Unit): Unit = block(this)
+
+  public companion object {
+    internal fun <T : EnumValue<T>> of(value: T): EnumLiteral<T> = EnumLiteral(value)
+  }
+}
