+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/FloatCache.kt
@@ -22,20 +22,3 @@ internal class FloatCache<T>(val init: (Float) -> T) {
     internal fun Int.isSmallInt() = this in 0..<SIZE
   }
 }
-
-internal class IntCache<T>(val init: (Int) -> T) {
-  private val smallInts = List(SIZE) { init(it) }
-
-  operator fun get(int: Int): T {
-    return when {
-      int.isSmallInt() -> smallInts[int]
-      else -> init(int)
-    }
-  }
-
-  companion object {
-    const val SIZE = 512
-
-    internal fun Int.isSmallInt() = this in 0..<SIZE
-  }
-}
