+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/FunctionCall.kt
@@ -1,0 +1,28 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+import dev.sargunv.maplibrecompose.expression.ExpressionContext
+import dev.sargunv.maplibrecompose.expression.value.ExpressionValue
+
+/** An [Expression] representing a function call. */
+public data class FunctionCall
+private constructor(
+  val name: String,
+  val args: List<Expression<*>>,
+  val isLiteralArg: (Int) -> Boolean,
+) : Expression<ExpressionValue> {
+  override fun compile(context: ExpressionContext): CompiledExpression<ExpressionValue> =
+    CompiledFunctionCall.of(name, args.map { it.compile(context) }, isLiteralArg)
+
+  override fun visit(block: (Expression<*>) -> Unit) {
+    block(this)
+    args.forEach { it.visit(block) }
+  }
+
+  public companion object {
+    public fun of(
+      name: String,
+      vararg args: Expression<*>,
+      isLiteralArg: (Int) -> Boolean = { false },
+    ): FunctionCall = FunctionCall(name, args.asList(), isLiteralArg)
+  }
+}
