+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/NullLiteral.kt
@@ -1,0 +1,8 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+/** A [Literal] representing a `null` value. */
+public data object NullLiteral : CompiledLiteral<Nothing, Nothing?> {
+  override val value: Nothing? = null
+
+  override fun visit(block: (Expression<*>) -> Unit): Unit = block(this)
+}
