+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/OffsetLiteral.kt
@@ -1,0 +1,17 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+import androidx.compose.ui.geometry.Offset
+import dev.sargunv.maplibrecompose.expression.value.FloatOffsetValue
+
+/** A [Literal] representing a [Offset] value. */
+public data class OffsetLiteral private constructor(override val value: Offset) :
+  CompiledLiteral<FloatOffsetValue, Offset> {
+  override fun visit(block: (Expression<*>) -> Unit): Unit = block(this)
+
+  public companion object {
+    private val zero = OffsetLiteral(Offset.Zero)
+
+    public fun of(value: Offset): OffsetLiteral =
+      if (value == Offset.Zero) zero else OffsetLiteral(value)
+  }
+}
