+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/ast/Options.kt
@@ -1,0 +1,23 @@
+package dev.sargunv.maplibrecompose.expression.ast
+
+import dev.sargunv.maplibrecompose.expression.ExpressionContext
+import dev.sargunv.maplibrecompose.expression.value.ExpressionValue
+import dev.sargunv.maplibrecompose.expression.value.MapValue
+
+/** An [Expression] representing a JSON object with values all [Expression]. */
+public data class Options<T : ExpressionValue>
+private constructor(val value: Map<String, Expression<T>>) : Expression<MapValue<T>> {
+
+  override fun compile(context: ExpressionContext): CompiledOptions<T> =
+    CompiledOptions.of(value.mapValues { it.value.compile(context) })
+
+  override fun visit(block: (Expression<*>) -> Unit) {
+    block(this)
+    value.values.forEach { it.visit(block) }
+  }
+
+  public companion object {
+    internal fun build(block: MutableMap<String, Expression<*>>.() -> Unit) =
+      Options(mutableMapOf<String, Expression<*>>().apply(block).mapValues { it.value })
+  }
+}
