+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/CirclePitchAlignment.kt
@@ -1,0 +1,13 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+
+/** Orientation of circles when the map is pitched. */
+public enum class CirclePitchAlignment(override val literal: StringLiteral) :
+  EnumValue<CirclePitchAlignment> {
+  /** Circles are aligned to the plane of the map, i.e. flat on top of the map. */
+  Map(StringLiteral.of("map")),
+
+  /** Circles are aligned to the plane of the viewport, i.e. facing the camera. */
+  Viewport(StringLiteral.of("viewport")),
+}
