+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/ExpressionType.kt
@@ -1,0 +1,14 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+import dev.sargunv.maplibrecompose.expression.dsl.type
+
+/** The type of value resolved from an expression, as returned by [type]. */
+public enum class ExpressionType(override val literal: StringLiteral) : EnumValue<ExpressionType> {
+  Number(StringLiteral.of("number")),
+  String(StringLiteral.of("string")),
+  Object(StringLiteral.of("object")),
+  Boolean(StringLiteral.of("boolean")),
+  Color(StringLiteral.of("color")),
+  Array(StringLiteral.of("array")),
+}
