+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/GeometryType.kt
@@ -1,0 +1,13 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+
+/** Type of a GeoJson feature, as returned by [Feature.type]. */
+public enum class GeometryType(override val literal: StringLiteral) : EnumValue<GeometryType> {
+  Point(StringLiteral.of("Point")),
+  LineString(StringLiteral.of("LineString")),
+  Polygon(StringLiteral.of("Polygon")),
+  MultiPoint(StringLiteral.of("MultiPoint")),
+  MultiLineString(StringLiteral.of("MultiLineString")),
+  MultiPolygon(StringLiteral.of("MultiPolygon")),
+}
