+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/IlluminationAnchor.kt
@@ -1,0 +1,14 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+
+/** Direction of light source when map is rotated. */
+public enum class IlluminationAnchor(override val literal: StringLiteral) :
+  EnumValue<IlluminationAnchor> {
+
+  /** The hillshade illumination is relative to the north direction. */
+  Map(StringLiteral.of("map")),
+
+  /** The hillshade illumination is relative to the top of the viewport. */
+  Viewport(StringLiteral.of("viewport")),
+}
