+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/LineJoin.kt
@@ -1,0 +1,24 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+
+/** Display of joined lines */
+public enum class LineJoin(override val literal: StringLiteral) : EnumValue<LineJoin> {
+  /**
+   * A join with a squared-off end which is drawn beyond the endpoint of the line at a distance of
+   * one-half of the line's width.
+   */
+  Bevel(StringLiteral.of("bevel")),
+
+  /**
+   * A join with a rounded end which is drawn beyond the endpoint of the line at a radius of
+   * one-half of the line's width and centered on the endpoint of the line.
+   */
+  Round(StringLiteral.of("round")),
+
+  /**
+   * A join with a sharp, angled corner which is drawn with the outer sides beyond the endpoint of
+   * the path until they meet.
+   */
+  Miter(StringLiteral.of("miter")),
+}
