+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/SymbolZOrder.kt
@@ -1,0 +1,30 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+
+/**
+ * Determines whether overlapping symbols in the same layer are rendered in the order that they
+ * appear in the data source or by their y-position relative to the viewport. To control the order
+ * and prioritization of symbols otherwise, use `sortKey`.
+ */
+public enum class SymbolZOrder(override val literal: StringLiteral) : EnumValue<SymbolZOrder> {
+  /**
+   * Sorts symbols by `sortKey` if set. Otherwise, sorts symbols by their y-position relative to the
+   * viewport if `iconAllowOverlap` or `textAllowOverlap` is set to `true` or `iconIgnorePlacement`
+   * or `textIgnorePlacement` is `false`.
+   */
+  Auto(StringLiteral.of("auto")),
+
+  /**
+   * Sorts symbols by their y-position relative to the viewport if `iconAllowOverlap` or
+   * `textAllowOverlap` is set to `true` or `iconIgnorePlacement` or `textIgnorePlacement` is
+   * `false`.
+   */
+  ViewportY(StringLiteral.of("viewport-y")),
+
+  /**
+   * Sorts symbols by `sortKey` if set. Otherwise, no sorting is applied; symbols are rendered in
+   * the same order as the source data.
+   */
+  Source(StringLiteral.of("source")),
+}
