+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/TextJustify.kt
@@ -1,0 +1,18 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+
+/** Text justification options. */
+public enum class TextJustify(override val literal: StringLiteral) : EnumValue<TextJustify> {
+  /** The text is aligned towards the anchor position. */
+  Auto(StringLiteral.of("auto")),
+
+  /** The text is aligned to the left. */
+  Left(StringLiteral.of("left")),
+
+  /** The text is centered. */
+  Center(StringLiteral.of("center")),
+
+  /** The text is aligned to the right. */
+  Right(StringLiteral.of("right")),
+}
