+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/TextRotationAlignment.kt
@@ -1,0 +1,37 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+
+/**
+ * In combination with [SymbolPlacement], determines the rotation behavior of the individual glyphs
+ * forming the text.
+ */
+public enum class TextRotationAlignment(override val literal: StringLiteral) :
+  EnumValue<TextRotationAlignment> {
+  /**
+   * For [SymbolPlacement.Point], aligns text east-west. Otherwise, aligns text x-axes with the
+   * line.
+   */
+  Map(StringLiteral.of("map")),
+
+  /**
+   * Produces glyphs whose x-axes are aligned with the x-axis of the viewport, regardless of the
+   * [SymbolPlacement].
+   */
+  Viewport(StringLiteral.of("viewport")),
+
+  /**
+   * For [SymbolPlacement.Point], this is equivalent to [TextRotationAlignment.Viewport]. Otherwise,
+   * aligns glyphs to the x-axis of the viewport and places them along the line.
+   *
+   * **Note**: This value not supported on native platforms yet, see
+   * [maplibre-native#250](https://github.com/maplibre/maplibre-native/issues/250)**
+   */
+  ViewportGlyph(StringLiteral.of("viewport-glyph")),
+
+  /**
+   * For [SymbolPlacement.Point], this is equivalent to [TextRotationAlignment.Viewport]. Otherwise,
+   * this is equivalent to [TextRotationAlignment.Map].
+   */
+  Auto(StringLiteral.of("auto")),
+}
