+++ lib/maplibre-compose-expressions/src/commonMain/kotlin/dev/sargunv/maplibrecompose/expression/value/TranslateAnchor.kt
@@ -1,0 +1,13 @@
+package dev.sargunv.maplibrecompose.expression.value
+
+import dev.sargunv.maplibrecompose.expression.ast.StringLiteral
+
+/** Frame of reference for offsetting geometry. */
+public enum class TranslateAnchor(override val literal: StringLiteral) :
+  EnumValue<TranslateAnchor> {
+  /** Offset is relative to the map */
+  Map(StringLiteral.of("map")),
+
+  /** Offset is relative to the viewport */
+  Viewport(StringLiteral.of("viewport")),
+}
